Throw error if bias scan file type not supported by mattak
If the user loads a bias scan file that is of a raw daq data type,
but mattak was not built with librno-g support, creating the voltage
calibration device quietly succeeded.
But when the calibrated method is later called, generated a segfault.
This now errors out when librno-g support is required but not provided.