add radiantSampleRate() and radiantReadoutDelays() methods to mattak::Dataset
These are needed for the pyroot backend to not be super slow when creating EventInfos. Basically, we add individual readers for each of these (combining the two would actually make sense, but that requires a bit more code rework). Incidentally, we move some of the code for loading incomplete waveforms to one place and also expose currentEntry()